chaged the script folder
