+++ Project/ProjectSettings/EditorBuildSettings.asset
@@ -9,7 +9,7 @@ EditorBuildSettings:
     path: Assets/Scenes/MainMenu.unity
     guid: e3aa7abd5d3ba8749b03896274dfc29f
   - enabled: 1
-    path: Assets/Scenes/Desert.unity
+    path: Assets/Scenes/Desert-Rename1.unity
     guid: 9fc0d4010bbf28b4594072e72b8655ab
   - enabled: 1
     path: Assets/Scenes/GameOver.unity
